FFT cleanup
