+++ Common/Common.vcxproj
@@ -391,13 +391,11 @@
     <ClInclude Include="..\ext\at3_standalone\atrac3plus.h" />
     <ClInclude Include="..\ext\at3_standalone\atrac3plus_data.h" />
     <ClInclude Include="..\ext\at3_standalone\avcodec.h" />
-    <ClInclude Include="..\ext\at3_standalone\bytestream.h" />
     <ClInclude Include="..\ext\at3_standalone\channel_layout.h" />
     <ClInclude Include="..\ext\at3_standalone\compat.h" />
     <ClInclude Include="..\ext\at3_standalone\fft.h" />
     <ClInclude Include="..\ext\at3_standalone\float_dsp.h" />
     <ClInclude Include="..\ext\at3_standalone\get_bits.h" />
-    <ClInclude Include="..\ext\at3_standalone\intfloat.h" />
     <ClInclude Include="..\ext\at3_standalone\intmath.h" />
     <ClInclude Include="..\ext\at3_standalone\intreadwrite.h" />
     <ClInclude Include="..\ext\at3_standalone\log.h" />
@@ -611,7 +609,7 @@
     <ClCompile Include="..\ext\at3_standalone\get_bits.c" />
     <ClCompile Include="..\ext\at3_standalone\channel_layout.c" />
     <ClCompile Include="..\ext\at3_standalone\compat.c" />
-    <ClCompile Include="..\ext\at3_standalone\fft_template.c" />
+    <ClCompile Include="..\ext\at3_standalone\fft.c" />
     <ClCompile Include="..\ext\at3_standalone\float_dsp.c" />
     <ClCompile Include="..\ext\at3_standalone\mathematics.c" />
     <ClCompile Include="..\ext\at3_standalone\mdct_template.c" />
