+++ ext/at3_standalone/channel_layout.c
@@ -24,6 +24,8 @@
  */
 
 #include <stdint.h>
+#include <string.h>
+#include <stdlib.h>
 #include <errno.h>
 #include <stddef.h>
