+++ ext/at3_standalone/fft.h
@@ -39,6 +39,8 @@
 
 #include <stdint.h>
 
+#include "compat.h"
+
 typedef float FFTSample;
 
 typedef struct FFTComplex {
@@ -46,8 +48,6 @@ typedef struct FFTComplex {
 } FFTComplex;
 
 typedef struct FFTContext FFTContext;
-
-#define FFT_NAME(x) x
 
 typedef float FFTDouble;
 
@@ -88,14 +88,8 @@ struct FFTContext {
     enum mdct_permutation_type mdct_permutation;
 };
 
-#if CONFIG_HARDCODED_TABLES
-#define COSTABLE_CONST const
-#else
-#define COSTABLE_CONST
-#endif
-
 #define COSTABLE(size) \
-    COSTABLE_CONST DECLARE_ALIGNED(32, FFTSample, FFT_NAME(ff_cos_##size))[size/2]
+     DECLARE_ALIGNED(32, FFTSample, ff_cos_##size)[size/2]
 
 extern COSTABLE(16);
 extern COSTABLE(32);
@@ -110,18 +104,13 @@ extern COSTABLE(8192);
 extern COSTABLE(16384);
 extern COSTABLE(32768);
 extern COSTABLE(65536);
-extern COSTABLE_CONST FFTSample* const FFT_NAME(ff_cos_tabs)[17];
-
-#define ff_init_ff_cos_tabs FFT_NAME(ff_init_ff_cos_tabs)
+extern FFTSample* const ff_cos_tabs[17];
 
 /**
  * Initialize the cosine table in ff_cos_tabs[index]
  * @param index index in ff_cos_tabs array of the table to initialize
  */
 void ff_init_ff_cos_tabs(int index);
-
-#define ff_fft_init FFT_NAME(ff_fft_init)
-#define ff_fft_end  FFT_NAME(ff_fft_end)
 
 /**
  * Set up a complex FFT.
@@ -131,9 +120,6 @@ void ff_init_ff_cos_tabs(int index);
 int ff_fft_init(FFTContext *s, int nbits, int inverse);
 
 void ff_fft_end(FFTContext *s);
-
-#define ff_mdct_init FFT_NAME(ff_mdct_init)
-#define ff_mdct_end  FFT_NAME(ff_mdct_end)
 
 int ff_mdct_init(FFTContext *s, int nbits, int inverse, double scale);
 void ff_mdct_end(FFTContext *s);
